+++ Frameworks/Navigine.xcframework/ios-arm64/Navigine.framework/Headers/NCRectangle.h
@@ -11,6 +11,15 @@
  *
  * @brief A rectangle with specified origin and size.
  *
+ *
+ *
+ * Swift code snippet:
+ * @snippet BitmapRegionDecoderExample.swift swift_Rectangle_constructor
+ *
+ * Objective C code snippet:
+ * @snippet BitmapRegionDecoderExample.m objc_Rectangle_constructor
+ *
+ *
  */
 DEFAULT_EXPORT_ATTRIBUTE
 @interface NCRectangle : NSObject
@@ -18,45 +27,44 @@ DEFAULT_EXPORT_ATTRIBUTE
 /** 
  * @brief Default constructor for class NCRectangle 
  */
-- (nonnull instancetype)initWithX:(float)x
-                                y:(float)y
-                            width:(float)width
-                           height:(float)height;
+- (nonnull instancetype)initWithX:(int32_t)x
+                                y:(int32_t)y
+                            width:(int32_t)width
+                           height:(int32_t)height;
 
 /** 
  * @brief Factory method for class NCRectangle 
  */
-+ (nonnull instancetype)rectangleWithX:(float)x
-                                     y:(float)y
-                                 width:(float)width
-                                height:(float)height;
++ (nonnull instancetype)rectangleWithX:(int32_t)x
+                                     y:(int32_t)y
+                                 width:(int32_t)width
+                                height:(int32_t)height;
 
 /**
  *
  * @brief rectangle's origin `x` coordinate.
  *
  */
-@property (nonatomic, readonly) float x;
+@property (nonatomic, readonly) int32_t x;
 
 /**
  *
  * @brief rectangle's origin `y` coordinate.
- *
  */
-@property (nonatomic, readonly) float y;
+@property (nonatomic, readonly) int32_t y;
 
 /**
  *
  * @brief rectangle's width.
  *
  */
-@property (nonatomic, readonly) float width;
+@property (nonatomic, readonly) int32_t width;
 
 /**
  *
  * @brief rectangle's height.
  *
  */
-@property (nonatomic, readonly) float height;
+@property (nonatomic, readonly) int32_t height;
 
 @end
